Move alert contact into alert_contact.go
