+++ pkg/alert_contact.go
@@ -1,0 +1,21 @@
+package uptimerobot
+
+// AlertContact represents an alert contact.
+type AlertContact struct {
+	ID           string `json:"id"`
+	FriendlyName string `json:"friendly_name"`
+	Type         int    `json:"type"`
+	Status       int    `json:"status"`
+	Value        string `json:"value"`
+}
+
+const alertContactTemplate = `ID: {{ .ID }}
+Name: {{ .FriendlyName }}
+Type: {{ .Type }}
+Status: {{ .Status }}
+Value: {{ .Value }}`
+
+// String returns a pretty-printed version of the alert contact.
+func (a AlertContact) String() string {
+	return render(alertContactTemplate, a)
+}
